clang

+++ src/kernel/gmp++/gmp++_int_rand.inl
@@ -48,7 +48,7 @@ namespace Givaro {
 
 	void Integer::seeding()
 	{
-		Integer::seeding( BaseTimer::seed() );
+		Integer::seeding( (unsigned long) BaseTimer::seed() );
 	}
